refactor: ページの責務分担を代表画面で揃える
フォーム一覧とユーザー管理で page、PageContent、View の境界を明示し、新しい画面追加時に参照しやすい構成にする。
